+++ test/learningAgentTest.cpp
@@ -170,6 +170,34 @@ TEST_F(LearningAgentTest, IsRootEvalSkipped)
            "maxNbEvaluationPerPolicy.";
 }
 
+TEST_F(LearningAgentTest, MakeJob)
+{
+    Learn::LearningAgent la(le, set, params);
+    la.init();
+    auto job = *la.makeJob(0);
+    ASSERT_EQ(1,job.getSize())
+    <<"job made in LearningAgent makeJob should contain 1 root only";
+    ASSERT_EQ(la.getTPGGraph().getRootVertices().at(0),job[0])
+                                <<"Encapsulate the root in a job shouldn't change it";
+
+}
+
+TEST_F(LearningAgentTest, MakeJobs)
+{
+    Learn::LearningAgent la(le, set, params);
+    la.init();
+    auto jobs = la.makeJobs(Learn::TRAINING);
+    ASSERT_EQ(la.getTPGGraph().getNbRootVertices(),jobs.size())
+                                <<"There should be as many jobs as roots";
+    for(int i=0; i<la.getTPGGraph().getNbRootVertices(); i++) {
+        ASSERT_EQ(1, (*jobs.front()).getSize())
+                                    << "job made in LearningAgent makeJobs should contain 1 root only";
+        ASSERT_EQ(la.getTPGGraph().getRootVertices().at(i), (*jobs.front())[0])
+                                    << "Encapsulate the root in a job shouldn't change it";
+        jobs.pop();
+    }
+}
+
 TEST_F(LearningAgentTest, EvalRoot)
 {
     params.archiveSize = 50;
@@ -186,8 +214,6 @@ TEST_F(LearningAgentTest, EvalRoot)
     la.init();
     std::shared_ptr<Learn::EvaluationResult> result;
     auto job = *la.makeJob(0);
-    ASSERT_EQ(la.getTPGGraph().getRootVertices().at(0),job[0])
-    <<"Encapsulate the root in a job shouldn't change it";
     ASSERT_NO_THROW(
         result = la.evaluateJob(tee, job,
                                 0, Learn::LearningMode::TRAINING, le))