+++ Changelog.md
@@ -6,6 +6,7 @@ _aaaa.mm.dd_
 ### New features
 * Fix code for building GEGELATI with clang (v7+) and added a dedicated job with travis-ci.
 * Add osx job in travis-ci.
+* Add adversarial training support. AdversarialLearningAgent can be declared and used with an AdversarialLearningEnvironment to train random groups of several roots together.
 
 ### Changes
 * Learning Agents now handle jobs instead of simple roots, allowing new concepts like adversarial learning (a job containing several roots that will train together).